backend/requirements.txt
fastapi
uvicorn
celery
redis
psycopg2-binary
python-dotenv
requests
google-generativeai>=0.8.3
python-jose[cryptography]
sqlalchemy
pydantic[email]
email-validator
passlib==1.7.4
bcrypt==3.2.2
python-multipart

